move config logic to utils path
